requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
jinja2
python-multipart
dbt-postgres
passlib[bcrypt]
bcrypt<4.0
itsdangerous
